Fix README.md
Second sentence: changed to be more natural (not sure about that, but better than what it was) and with less repetition
in `Building/Running` : changed Rename to Copy because if you rename a file, git will remove the `base.dub.json` file which is necessary for new contributors to get started. I intended `base.dub.json` to be a static file that doesn't change and for each contributor to have their own git-ignored `dub.json` that is specific to their system.